Add cmd/internal package, remove enableLogging from config file
